scala 2.13 + show extension
