+++ shared/src/main/scala/io/github/hamsters/Validation.scala
@@ -109,18 +109,4 @@ object Validation {
     */
   def hasFailures[L](eithers: Either[L, _]*): Boolean = failures(eithers: _*).nonEmpty
 
-  implicit class RightBiasedEither[L, R](e: Either[L, R]) {
-    def map[R2](f: R => R2) = e.right.map(f)
-
-    def flatMap[R2](f: R => Either[L, R2]) = e.right.flatMap(f)
-
-    def filter(p: (R) => Boolean) = filterWith(p)
-
-    def filterWith(p: (R) => Boolean) = e.right.filter(p)
-
-    def get = e.right.get
-
-    def getOrElse[R2 >: R](or: => R2) = e.right.getOrElse(or)
-  }
-
 }
